LayoutTests/platform/ios/media/modern-media-controls/compact-media-controls/compact-media-controls-layout-expected.txt

CONSOLE MESSAGE: line 140: Button failed to load, iconName = PlayCompact, layoutTraits = 8, src = PlayCompact.pdf
CONSOLE MESSAGE: line 140: Button failed to load, iconName = InvalidCompact, layoutTraits = 8, src = InvalidCompact.pdf
CONSOLE MESSAGE: line 140: Button failed to load, iconName = ActivityIndicatorSpriteCompact, layoutTraits = 8, src = ActivityIndicatorSpriteCompact@2x.png
Testing the CompactMediaControls layout.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


With controls at their default state (paused) only the play button is showing.
PASS mediaControls.children is [mediaControls.playButton]

With controls at pending state, only the activity indicator is showing.
PASS mediaControls.children is [mediaControls.activityIndicator]

With controls at invalid state, only the invalid icon is showing.
PASS mediaControls.children is [mediaControls.invalidButton]

PASS successfullyParsed is true

TEST COMPLETE
